remove CancelableStdin, simplify FillableStdin
I don't see the point of having reads to stdin be cancelable at this layer;
the main concern is to be able to unblock `(*Terminal).GetRune`, which
can just be handled at the `Terminal` layer instead.

Also add locking around transitions in and out of raw mode. This prevents
a data race, but is necessary because the underlying operation is inherently
race-y (and can be triggered asynchronously due to concurrent Readline() and
Close()).